cambios ultima reunion

+++ database/migrations/2025_07_23_220803_add_tipo_cortina_to_detalle_cotizacion_table.php
@@ -18,8 +18,10 @@ return new class extends Migration
 
     public function down()
     {
-        Schema::table('detalle_cotizacion', function ($table) {
-            $table->dropColumn('tipo_cortina');
-        });
+        if (Schema::hasColumn('detalle_cotizacion', 'tipo_cortina')) {
+            Schema::table('detalle_cotizacion', function ($table) {
+                $table->dropColumn('tipo_cortina');
+            });
+        }
     }
 };
